Share current pull request detection across frontends
