+++ packages/agent-repository/agent-repository.cabal
@@ -47,7 +47,8 @@ library
     exposed-modules:  Agent.CLI.RepositoryDelivery
                     , Agent.CLI.RepositoryDelivery.ConfirmationStore
                     , Agent.CLI.RepositoryReview
-    build-depends:    aeson >= 2.0
+    build-depends:    agent-cli-runtime >= 0.1 && < 0.2
+                    , aeson >= 2.0
                     , async
                     , base >= 4.17 && < 5
                     , bytestring
