+++ scripts/check-package-boundaries.sh
@@ -49,9 +49,10 @@ if rg --line-number '^import[[:space:]]+(qualified[[:space:]]+)?Agent\.(CLI|TUI)
   fail "frontend types leaked into the session lifecycle kernel"
 fi
 
-# Process resources retain legacy module names, but their implementations must
-# stay in the shared package. CLI facades may import them, not redefine them.
-for module_path in Agent/CLI/NativeProcess.hs Agent/CLI/Session/Threads.hs; do
+# Shared resources and session policy retain legacy module names, but their
+# implementations must stay in the runtime. CLI facades may import them,
+# not redefine them.
+for module_path in Agent/CLI/NativeProcess.hs Agent/CLI/Session/Threads.hs Agent/CLI/Session/PullRequest.hs; do
   [[ -f "$runtime/src/$module_path" ]] || fail "missing shared resource: $module_path"
   [[ ! -e "$cli/src/$module_path" ]] || fail "resource implementation returned to CLI: $module_path"
 done
@@ -110,7 +111,7 @@ if rg --line-number '\bagent-cli\b' \
   fail "agent-external-session must remain independent of agent-cli"
 fi
 
-if rg --line-number '\bagent-cli\b' \
+if rg --line-number '\bagent-cli([[:space:],><=]|$)' \
   "$repository/agent-repository.cabal"; then
   fail "agent-repository must remain independent of agent-cli"
 fi
